+++ stubs/Netwerkstatt/SilverstripeRector/Tests/Misc/SilverstripeDeprecationCommentRector/Source/ClassWithDeprecatedMethod.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Netwerkstatt\SilverstripeRector\Tests\Misc\SilverstripeDeprecationCommentRector\Source;
+
+class ClassWithDeprecatedMethod
+{
+    public function deprecatedMethod()
+    {
+    }
+}