+++ stubs/Netwerkstatt/SilverstripeRector/Tests/Misc/SilverstripeDeprecationCommentRector/Source/DeprecatedClass.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Netwerkstatt\SilverstripeRector\Tests\Misc\SilverstripeDeprecationCommentRector\Source;
+
+class DeprecatedClass
+{
+}